[Rename] 패키지 분리
